requirements.txt
alphacast
pandas
requests
python-dotenv
xlrd